backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pinecone
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-core
pypdf
python-dotenv
# PostgreSQL and Database
psycopg2-binary
sqlalchemy
bcrypt
alembic
# Payment Gateway and Subscriptions
razorpay==1.4.1
python-dateutil==2.8.2
reportlab==4.0.7
